Replacing int with long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) throws Exception {
 
-        int[] sales1 = {1, 2, 5, 586, 0, 56, -8};
+        long[] sales1 = {1, 2, 5, 586, 0, 56, -8};
         SalesManager manager1 = new SalesManager(sales1);
 
         System.out.println("Максимальная продажа - " + manager1.max());
